Create README.md
This is a web-based game where there is a randomly generated number set as the goal for the player to reach.
There are four buttons gem images on them, and each button has a random number assigned to them and the player must add them up to match the goal.
If the player goes over the goal they lose, they must match their number to the goal number. 
The player doesn't know the number assigned to each button, they have to calculate it from the "number counter" that adds up the 
numbers of each button pressed.
Everytime the player wins or loses the game resets.